Avoid rebooting twice when installing WSL
For some reason, after upgrading to WiX v5, the forceReboot action
appeared before the Dism actions in the msi InstallExecuteSequence
table. As a consequence the user was asked to reboot before WSL or
Hyper-V installation and to reboot a second time after their
installation.

To avoid that the ForceReboot action field `Before=StopServices` is
replaced with `After=DismX86`

Signed-off-by: Mario Loriedo <mario.loriedo@gmail.com>