archive: mark some places which break zero-copy
The Go standard library supports copy_file_range for I/O between
files. For most file systems, this is nice speed up, because it avoids
copying data.  For certain filesystems, like BRTFS and XFS, CoW causes
such writes to be metadata-only, speeding them up by 10x or so.

In principle, uncompressed tar files can be read and written using
copy_file_range, as the file format does not checksum the data (see
https://github.com/golang/go/issues/70807). Currently, podman makes
this impossible, as the archive package liberally inserts buffers and
pipes.

This commit marks a couple of places that should be revisited.