Avoid indirect links through quadlet(5)
On Linux systems, the quadlet(5) manpage points to the actual content at
podman-systemd.unit(5) but this cannot be counted on elsewhere.  In
particular, this symlink isn't installed by the macOS Brew package, and
https://docs.podman.io/en/latest/markdown/quadlet.5.html is a broken
URL.  Symlinks are also unlikely to function properly within the Windows
distribution, though this is untested speculation.

Now that an HTML link to podman-systemd.unit.5.html can be counted on to
work properly, this change also adds hyperlinks to these references.

Signed-off-by: Warren Young <wyoung@tangentsoft.com>