remove bufio from container stdin
When we read the stdin to forward it into the container it doesn't
really make sense to have it buffered. In fact our detach.Copy() is
already reading in large buffers so we do not gain anything by it.

In fact this is causing a bug with the new detach.Copy() code from
c/common. The Copy() function now calls into the std io.Copy() function
if no detach keys are given. That function tries to use the WriterTo()
method if available. bufio.Buffer has that and this function causes a
empty write to the conmon attach socket[1]. I think this is a bug in the
standard library.
Now because the conmon socket where it writes to is a packet based
socket and not a stream one the reading side (conmon) read an empty
packet. This makes it think we hit EOF, not sure if this is intentional
over there either.

To avoid these bugs stop using bufio here. Now we keep using it in the
bindings as the API there is stable so I cannot just change the type
there and on the remote side it should not cause any issues as we do not
write to conmon there.

Now I did saw commit 1df4dba0a0 ("Switch to bufio Reader for exec streams")
but I don't quite follow the reason there. The code has changed a fair
bit since then so I don't think the races should still be there. If they
are the CI tests should show them I presume.

[1] https://cs.opensource.google/go/go/+/refs/tags/go1.23.5:src/bufio/bufio.go;l=518-521

Signed-off-by: Paul Holzinger <pholzing@redhat.com>