[v5.2-rhel] Fix HealthCheck log destination, count, and size defaults
GoLang sets unset values to the default value of the type. This means that the destination of the log is an empty string and the count and size are set to 0. However, this means that size and count are unbounded, and this is not the default behavior.

Fixes: https://github.com/containers/podman/issues/25473
Fixes: https://issues.redhat.com/browse/RHEL-83558

Signed-off-by: Jan Rodák <hony.com@seznam.cz>
(cherry picked from commit fff42ac232ecdece93b0ea6053a84f5c3318ea72)
Signed-off-by: Jan Rodák <hony.com@seznam.cz>