Verify images in manifest list are not dangling and pruned
Vendor in latest containers/common

Currently if you create an un tagged image and add it to a manifest
list, podman image prune will remove the image, and leave you with
a broken manifest list. This PR removes the image from dangling in
this situation and prevents the pruning.

We have seen this trigger issues with RamaLama which is creating
manifest lists in this manner, and then users pruning the images.

Verifing: https://github.com/containers/common/pull/2360

Signed-off-by: Daniel J Walsh <dwalsh@redhat.com>