remove github.com/crc-org/crc/v2 dependency
There is one user in the code and one in the tests.
First the test is testing an external function that doesn't seem to make
a lot of sense.

Second when we pull the machine image and renmae doesn't work it would
mean it is on a different disk. That should never happen.
The oci pull is to
~/.local/share/containers/podman/machine/<provider>/cache/<sha>/...
and then it gets moved to
~/.local/share/containers/podman/machine/qemu/cache/<sha>.<extension>

These can under normal circumstances never be on different disk as it
would mean the pulled sha is known ahead of time and then that the blob
path is a mount point which is unrealistic and not something we have to
support.

It also removes three transitive dependencies.

Signed-off-by: Paul Holzinger <pholzing@redhat.com>