Replace podman pause image with rootfs.
This commit removes the code to build a local pause
image from the Containerfile and keeps just a code
to PullInfraImage.

In case the user-defined infra image is not set, the
PrepareInfraSpec is called to find out the catatonit
binary and use it using the rootfs with overlay.

This removes the need to build a local pause container
image.

The same logic is also applied to createServiceContainer
which is originally also based on the pause image.

Fixes: #23292

Signed-off-by: Jan Kaluza <jkaluza@redhat.com>