cirrus: save/restore golangci-lint cache for linux
This shaves off some time from golangci-lint run.

For linux, it's down from 5 to 3 minutes (for the whole task, not just
the golangci-lint runs), and the cache is only ~7MB (according to
golangci-lint, probably uncompressed) or ~2MB (according to cirrus CI,
probably compressed).

For Mac, due to remote build tag we don't lint much so linting takes
less than a minute, so it is not worth optimizing.

For Windows, similar.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>